initiated Linear Algebra
