+++ GeometrySharp.sln
@@ -23,8 +23,8 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|x64.ActiveCfg = Debug|x64
-		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|x64.Build.0 = Debug|x64
+		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|x64.Build.0 = Debug|Any CPU
 		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Debug|x86.ActiveCfg = Debug|Any CPU
 		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3249FF54-061F-4A9B-8DFE-F69A2E5CFCBA}.Release|Any CPU.Build.0 = Release|Any CPU
