+++ GeometrySharp/Evaluation/Make.cs
@@ -28,7 +28,7 @@ namespace GeometrySharp.Evaluation
             knots.Add(lsum);
             var weights = points.Select(x => 1.0).ToList();
             points.ForEach(x => weights.Add(1.0));
-            return new NurbsCurve(1, knots, Eval.Homogenize1d(points, weights));
+            return new NurbsCurve(1, knots, LinearAlgebra.Homogenize1d(points, weights));
         }
 
         public static NurbsCurve RationalBezierCurve(List<Vector3> controlPoints, List<double> weights = null)
@@ -41,7 +41,7 @@ namespace GeometrySharp.Evaluation
                 knots.Add(1.0);
             if (weights == null)
                 weights = Sets.RepeatData(1.0, controlPoints.Count);
-            return new NurbsCurve(degree, knots, Eval.Homogenize1d(controlPoints, weights));
+            return new NurbsCurve(degree, knots, LinearAlgebra.Homogenize1d(controlPoints, weights));
             //weights = Sets.RepeatData(1.0, controlPoints.Count);
             //return new NurbsCurveData(degree, knots, Eval.Homogenize1d(controlPoints, weights));
             //return null;
